+++ src/main/java/com/clone/ohouse/store/domain/product/Product.java
@@ -24,8 +24,8 @@ public class Product {
 
     @Column(length = 45)
     private String productName;
-    private Integer price;
-    private Integer stock;
+    private Long price;
+    private Long stock;
     private Integer rateDiscount;
 
     private Long popular = 0L;
@@ -35,7 +35,7 @@ public class Product {
     private StorePosts storePosts;
 
     @Builder
-    public Product(Item item, String productName, Integer price, Integer stock, Integer rateDiscount, Long popular, StorePosts storePosts) {
+    public Product(Item item, String productName, Long price, Long stock, Integer rateDiscount, Long popular, StorePosts storePosts) {
         this.item = item;
         this.productName = productName;
         this.price = price;
@@ -49,7 +49,7 @@ public class Product {
         this.popular++;
     }
 
-    public void update(Item item, String productName, Integer stock, Integer price, Integer rateDiscount,StorePosts storePosts) {
+    public void update(Item item, String productName, Long stock, Long price, Integer rateDiscount,StorePosts storePosts) {
         if(item != null) this.item = item;
         if(productName != null) this.productName = productName;
         if(price != null) this.price = price;
@@ -60,12 +60,12 @@ public class Product {
 
     }
 
-    public void returnAmount(Integer count) throws Exception{
-        if(count <= 0) throw new RuntimeException("물품의 0이하의 수량을 되돌려 줄 수 없습니다.");
+    public void returnAmount(Long count) throws Exception{
+        if(count <= 0L) throw new RuntimeException("물품의 0이하의 수량을 되돌려 줄 수 없습니다.");
         this.stock += count;
     }
 
-    public OrderedProduct makeOrderedProduct(Order order, Integer price, Integer amount) throws Exception{
+    public OrderedProduct makeOrderedProduct(Order order, Long price, Long amount) throws Exception{
         if(amount > this.stock) throw new RuntimeException("재고보다 많이 주문할 수 없습니다");
 
         this.stock -= amount;
