+++ src/main/java/com/clone/ohouse/store/domain/storeposts/dto/StorePostsViewDto.java
@@ -18,10 +18,10 @@ public class StorePostsViewDto {
     private String title;
     private String previewImageUrl;
     private Long popular;
-    private Integer price;
+    private Long price;
     private Integer discountRate;
 
-    public StorePostsViewDto(StorePosts postEntity, Integer price, Integer discountRate, Long popular) {
+    public StorePostsViewDto(StorePosts postEntity, Long price, Integer discountRate, Long popular) {
         this.id = postEntity.getId();
         this.title = postEntity.getTitle();
         this.previewImageUrl = postEntity.getPreviewImageUrl();
